Extract "adc_handler" in separate module.
