+++ src/adc.c
@@ -64,40 +64,15 @@ static void MX_ADC1_Init(void);
 static void MX_ADC2_Init(void);
 static void MX_ADC3_Init(void);
 
-static int get_adc_value(enum AnalogPins name);
-
 /**************************************************************************************
  * FUNCTION DEFINITION
  **************************************************************************************/
-
-int adc_handler(uint8_t opcode, uint8_t *data, uint16_t size)
-{
-  if (opcode == CONFIGURE)
-  {
-    /* Note: ADC currently only supports polling mode.
-     * uint16_t adc_sample_rate = *((uint16_t*)data);
-     * dbg_printf("Setting ADC samplerate to %d milliseconds\n", adc_sample_rate);
-     */
-    return 0;
-  }
-  else if ((opcode >= A0) && (opcode <= A7))
-  {
-    return get_adc_value(opcode);
-  }
-  else
-  {
-    dbg_printf("adc_handler: invalid ADC opcode %02x\n", opcode);
-    return 0;
-  }
-}
 
 void adc_init() {
 
   MX_ADC1_Init();
   MX_ADC2_Init();
   MX_ADC3_Init();
-
-  register_peripheral_callback(PERIPH_ADC, &adc_handler);
 }
 
 int get_adc_value(enum AnalogPins name) {
